// frontend/src/api/model/getFoodsForAdminResponseDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */
import type { FoodDetailDto } from './foodDetailDto';

export interface GetFoodsForAdminResponseDto {
  foods: FoodDetailDto[];
  page: number;
  count: number;
  total: number;
}